+++ src/framebuffer/screen_modes.h
@@ -98,7 +98,7 @@ typedef struct screen_mode {
    void                    (*flash)(struct screen_mode *screen, int mark);
    void               (*set_colour)(struct screen_mode *screen, colour_index_t index, int r, int g, int b);
    pixel_t            (*get_colour)(struct screen_mode *screen, uint8_t gcol);
-   pixel_t        (*nearest_colour)(struct screen_mode *screen, int r, int g, int b);
+   pixel_t        (*nearest_colour)(struct screen_mode *screen, uint8_t r, uint8_t g, uint8_t b);
    void           (*update_palette)(struct screen_mode *screen, int mark);
    void                (*set_pixel)(struct screen_mode *screen, int x, int y, pixel_t value);
    pixel_t             (*get_pixel)(struct screen_mode *screen, int x, int y);
